Run conveyor when hopper is emptying, revert drive mode switcher to constant
The plan is to store balls in the conveyor, so when we are emptying the
hopper we need to run the conveyor to empty those balls into the hopper.
I wanted to have the drive mode controlled by a SmartDashboard input
rather than a constant so you can change it on the fly, but it wasn't
working so I switched it back.